Add Dockerfile for Ruby application
A Dockerfile has been added to containerize our Ruby application. This
Dockerfile starts from a Ruby 3.2.2 base and installs all the necessary
dependencies such as build-essential, apt-utils, libpq-dev and nodejs.
It also sets the work directory to '/docker/app', adds current directory
files to it and exposes default port. Main application server is Puma,
but Rails can also be used as an alternative.