Update vcs.xml, cable.yml, and .iml files; remove line in sqldialects.xml
This commit includes updates to several configuration files. In vcs.xml, commit message inspections were configured to keep commit messages clear and concise. The cable.yml file was changed to use Redis as the adapter for any development environment, improving the efficiency of data fetching. Several .iml files were updated, mostly to adjust library version details. This keeps the project in sync with the latest, most stable versions of the libraries. In sqldialects.xml, a line mapping a file to 'GenericSQL' was removed. This optimizes SQL dialect detection, assuming the project-wide setting of 'PostgreSQL'.