Add session and user login functionalities
This commit introduces several functionalities needed for user handling. It includes a session helper for managing user login sessions (login, verify logged in users) to the app and the related sessions controller to handle new, create, and destroy actions for a session. Alongside, the commit adds a sessions view for user login and test cases for both login and session creation. Also, minor modifications are done via adding a .idea file and javascript for a collapsible menu feature. These adjustments establish the basis for handling user logins, authentication and thus, increase the app's security and user management efficiency.